Rename client to sdk
