+++ fission/sdk/types.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package client
+package sdk
 
 const (
 	FISSION_DEPLOYMENT_NAME_KEY = "fission-name"
